Wiring up multiple http requests in a pool
