+++ composer.json
@@ -13,6 +13,7 @@
         "php": "^7.4|^8.0",
         "alleyinteractive/composer-wordpress-autoloader": "^0.4",
         "alleyinteractive/wp-asset-manager": "^1.0",
+        "alleyinteractive/wp-concurrent-remote-requests": "^0.1.0",
         "dragonmantank/cron-expression": "^3.1",
         "fakerphp/faker": "^1.16",
         "filp/whoops": "^2.14",
