+++ src/mantle/http-client/composer.json
@@ -4,6 +4,7 @@
     "type": "project",
     "require": {
         "php": "^7.4|^8.0",
+        "alleyinteractive/wp-concurrent-remote-requests": "^0.1",
         "mantle-framework/support": "^0.5"
     },
     "extra": {
