+++ tests/testing/concerns/test-interacts-with-external-requests.php
@@ -2,7 +2,8 @@
 namespace Mantle\Tests\Testing\Concerns;
 
 use Mantle\Facade\Http;
-use Mantle\Http_Client\Http_Client;
+// todo: convert to facade.
+use Mantle\Http_Client\Factory;
 use Mantle\Testing\Mock_Http_Response;
 use Mantle\Testing\Framework_Test_Case;
 use Mantle\Testing\Mock_Http_Sequence;
@@ -134,7 +135,7 @@ class Test_Interacts_With_External_Requests extends Framework_Test_Case {
 				->push_status( 500 )
 		);
 
-		$http = new Http_Client();
+		$http = new Factory();
 
 		$this->assertEquals( 200, $http->get( 'https://example.com/sequence/' )->status() );
 		$this->assertEquals( 400, $http->get( 'https://example.com/sequence/' )->status() );
@@ -151,7 +152,7 @@ class Test_Interacts_With_External_Requests extends Framework_Test_Case {
 				->push_status( 400 )
 		);
 
-		$http = new Http_Client();
+		$http = new Factory();
 
 		$this->assertEquals( 200, $http->get( 'https://example.com/sequence/' )->status() );
 		$this->assertEquals( 400, $http->get( 'https://example.com/sequence/' )->status() );
@@ -166,7 +167,7 @@ class Test_Interacts_With_External_Requests extends Framework_Test_Case {
 				->when_empty( Mock_Http_Response::create()->with_status( 202 ) )
 		);
 
-		$http = new Http_Client();
+		$http = new Factory();
 
 		$this->assertEquals( 200, $http->get( 'https://example.com/sequence/' )->status() );
 		$this->assertEquals( 400, $http->get( 'https://example.com/sequence/' )->status() );
